# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Naturascan API",
        "description": "Naturascan API",
        "contact": {
            "name": "Gilles",
            "email": "gillesakakpo01@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/especes": {
            "get": {
                "tags": [
                    "Espece"
                ],
                "summary": "List all espece species",
                "operationId": "99b2fcfaa92d93625782773193fd6c29",
                "responses": {
                    "200": {
                        "description": "List all espece species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Espece"
                ],
                "summary": "Create a new espece species",
                "operationId": "97a2369003e032742d2182a8c1d1f210",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Espece"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new espece species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/especes/{id}": {
            "get": {
                "tags": [
                    "Espece"
                ],
                "summary": "Show espece species by ID",
                "operationId": "544724144b7bd94139f98cdeb8a349d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the espece species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show espece species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Espece"
                ],
                "summary": "Update espece species by ID",
                "operationId": "248c3a19b23c99b1b6c93386d049abf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the espece species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Espece"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update espece species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Espece"
                ],
                "summary": "Delete espece species by ID",
                "operationId": "1cba9f9ddf18f587055f1447c3664b64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the espece species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete espece species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/etapes": {
            "get": {
                "tags": [
                    "Etape"
                ],
                "summary": "List all etape species",
                "operationId": "32b161076bb55d01b82545aeeb3fd7fb",
                "responses": {
                    "200": {
                        "description": "List all etape species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Etape"
                ],
                "summary": "Create a new etape species",
                "operationId": "ca69a7a3c28782350a9369e966928276",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Etape"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new etape species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/etapes/{id}": {
            "get": {
                "tags": [
                    "Etape"
                ],
                "summary": "Show etape species by ID",
                "operationId": "27fe5bf0cb2d5b2a9bea3545e1ca8534",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the etape species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show etape species by ID"
                    },
                    "404": {
                        "description": "Etape species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Etape"
                ],
                "summary": "Update etape species by ID",
                "operationId": "01a6536ea5f4c693a59a3350ab4b5f9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the etape species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Etape"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update etape species by ID"
                    },
                    "404": {
                        "description": "Etape species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Etape"
                ],
                "summary": "Delete etape species by ID",
                "operationId": "19bae0da57a20eac32424860d72f837d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the etape species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete etape species by ID"
                    },
                    "404": {
                        "description": "Etape species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Espece": {
                "title": "Espece",
                "description": "Espece",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer"
                    },
                    "common_name": {
                        "title": "common_name",
                        "type": "string"
                    },
                    "scientific_name": {
                        "title": "scientific_name",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "type": "string"
                    },
                    "category_id": {
                        "title": "category_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Etape": {
                "title": "Etape",
                "description": "Etape",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer"
                    },
                    "nom": {
                        "title": "nom",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "type": "string"
                    },
                    "heure_depart_port": {
                        "title": "heure_depart_port",
                        "type": "string"
                    },
                    "heure_arrivee_port": {
                        "title": "heure_arrivee_port",
                        "type": "string"
                    },
                    "sortie_id": {
                        "title": "sortie_id",
                        "type": "integer"
                    },
                    "point_de_passage_id": {
                        "title": "point_de_passage_id",
                        "type": "integer"
                    },
                    "departure_weather_report_id": {
                        "title": "departure_weather_report_id",
                        "type": "integer"
                    },
                    "arrival_weather_report_id": {
                        "title": "arrival_weather_report_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}